--- a/app/javascript/packages/build-sass/package.json
+++ b/app/javascript/packages/build-sass/package.json
@@ -7,10 +7,9 @@
     "build-sass": "./cli.js"
   },
   "dependencies": {
-    "autoprefixer": "^10.4.4",
+    "@parcel/css": "^1.12.0",
+    "browserslist": "^4.21.2",
     "chokidar": "^3.5.3",
-    "cssnano": "^5.1.7",
-    "postcss": "^8.4.12",
     "sass-embedded": "^1.53.0"
   }
 }