**Why**:

- Improved performance, both at build time and runtime.
- Fewer dependencies to manage and maintain

I had tried this once before and the feature support with autoprefixing didn't have what we needed at the time (especially [CSS mask](https://developer.mozilla.org/en-US/docs/Web/CSS/mask#browser_compatibility) used by a lot of USWDS icons), but it looks like this has since improved. Will be worth testing a bit more exhaustively to check overall support.

I also want to confirm more closely that the browserlist configuration is being used as expected.

**Resources:**

- https://github.com/parcel-bundler/parcel-css

**Impact:**

`yarn build:css` (development bundling)

_Before:_ ~3.5s
_After:_ ~2.2s

`NODE_ENV yarn build:css` (production bundling with minification)

_Before:_ ~5.3s
_After:_ ~2.5s

`NODE_ENV yarn build:css && brotli-size app/assets/builds/application.css` (production stylesheet size)

_Before:_ 36.6kb
_After:_ 36.0kb