@aduth 2022-08-02
I gave this a thorough shake locally, spot-checking individual code differences (particularly on vendor prefixes like `-webkit-`), and went through critical flows across primary supported browsers.

Next I'm going to test this in my personal sandbox to make sure the binaries don't give us any trouble in deployment, but otherwise this is shaping up nicely to merge.